+++ src/apps/chifra/pkg/rpcClient/client.go
@@ -287,12 +287,12 @@ func DecodeHex(hex string) []byte {
 }
 
 // GetBlockZeroTs for some reason block zero does not return a timestamp, so we assign block one's ts minus 14 seconds
-func GetBlockZeroTs(chain string) (uint64, error) {
+func GetBlockZeroTs(chain string) (int64, error) {
 	ts := rpc.GetBlockTimestamp(chain, 0)
 	if ts == 0 {
 		ts = rpc.GetBlockTimestamp(chain, 1) - 13
 	}
-	return uint64(ts), nil
+	return ts, nil
 }
 
 func GetCodeAt(chain, addr string, bn uint64) ([]byte, error) {
