+++ src/apps/chifra/pkg/tslib/tots.go
@@ -7,19 +7,19 @@ import (
 )
 
 // FromBnToTs returns a chain-specific Linux timestamp given a block number
-func FromBnToTs(chain string, bn uint64) (uint64, error) {
+func FromBnToTs(chain string, bn uint64) (int64, error) {
 	ret, err := FromBn(chain, bn)
-	return uint64(ret.Ts), err
+	return int64(ret.Ts), err
 }
 
 // FromDateToTs returns a Linux timestamp given a date string (not chain-specific)
-func FromDateToTs(dateStr string) (uint64, error) {
+func FromDateToTs(dateStr string) (int64, error) {
 	str := toIsoDateStr(dateStr)
 	t, err := gostradamus.Parse(str, gostradamus.Iso8601)
 	if err != nil {
 		return 0, err
 	}
-	return uint64(t.UnixTimestamp()), nil
+	return t.UnixTimestamp(), nil
 }
 
 func toIsoDateStr(dateStr string) string {
